Ejemplos_presentados/ejemplo_completo/frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["Edgar Martinez <emartinez@cikume.com>"]

[tool.poetry.dependencies]
python = "^3.12"
nicegui = "^2.5.0"
python-decouple = "^3.8"
requests = "^2.32.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
